Addition of User Guide
I added the user guide into the program and im pretty happy with it but im probably going to change it because I think it might be too long and I dont think ill have enough time to complete everything on there. Im also not too sure if ill have enough time to complete it.