Original Data Frame
I tried creating the original data frame and I was having trouble because it was very tedious. But after I managed to type everything in, I checked to see if it ran but it didn't. I realized I didn't put the commas so I did. Then I tried to run it again but it didn't work and it had a lot of errors but I still need to figure out why.