Finalised code
I finally finished the code and decided to remove the function that I didn't know how to do and just replaced it with a function that was much more simpler. The reason I decided to make the UI even more advanced is because I lack the skillset to understand more complex code at the moment and no longer have time because I must spend time doing the evaluation.